Rename RequestsSender to RequestSender and move to package backend.api
