+++ app/src/test/java/com/rebalance/backend/api/RequestSenderTest.kt
@@ -1,15 +1,15 @@
 package com.rebalance.backend.api
 
 import com.google.gson.Gson
-import com.rebalance.backend.api.request.RequestsSender.Companion.sendDelete
-import com.rebalance.backend.api.request.RequestsSender.Companion.sendGet
-import com.rebalance.backend.api.request.RequestsSender.Companion.sendPost
+import com.rebalance.backend.api.RequestSender.Companion.sendDelete
+import com.rebalance.backend.api.RequestSender.Companion.sendGet
+import com.rebalance.backend.api.RequestSender.Companion.sendPost
 import com.rebalance.backend.api.entities.LoginAndPassword
 import com.rebalance.backend.exceptions.ServerException
 import org.junit.Assert.fail
 import org.junit.Test
 
-class RequestsSenderTest {
+class RequestSenderTest {
     private val ip = "http://192.168.180.253:8080/";
 
     @Test
